sensor/requirements.txt
googlemaps==2.5.1
Shapely==1.6.4.post1
kafka-python==1.4.2
